+++ src/main/java/io/strimzi/kafka/bridge/Application.java
@@ -5,10 +5,13 @@
 
 package io.strimzi.kafka.bridge;
 
-import io.micrometer.core.instrument.MeterRegistry;
 import io.strimzi.kafka.bridge.config.BridgeConfig;
 import io.strimzi.kafka.bridge.config.ConfigRetriever;
 import io.strimzi.kafka.bridge.http.HttpBridge;
+import io.strimzi.kafka.bridge.metrics.JmxCollectorRegistry;
+import io.strimzi.kafka.bridge.metrics.MetricsReporter;
+import io.strimzi.kafka.bridge.metrics.MetricsType;
+import io.strimzi.kafka.bridge.metrics.StrimziCollectorRegistry;
 import io.strimzi.kafka.bridge.tracing.TracingUtil;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
@@ -18,7 +21,6 @@ import io.vertx.micrometer.Label;
 import io.vertx.micrometer.MetricsDomain;
 import io.vertx.micrometer.MicrometerMetricsOptions;
 import io.vertx.micrometer.VertxPrometheusOptions;
-import io.vertx.micrometer.backends.BackendRegistries;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.Option;
@@ -34,6 +36,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.EnumSet;
 import java.util.HashSet;
 import java.util.Map;
@@ -46,8 +49,6 @@ import java.util.stream.Collectors;
 public class Application {
     private static final Logger LOGGER = LogManager.getLogger(Application.class);
 
-    private static final String KAFKA_BRIDGE_METRICS_ENABLED = "KAFKA_BRIDGE_METRICS_ENABLED";
-
     /**
      * Bridge entrypoint
      *
@@ -56,27 +57,12 @@ public class Application {
     public static void main(String[] args) {
         LOGGER.info("Strimzi Kafka Bridge {} is starting", Application.class.getPackage().getImplementationVersion());
         try {
-            VertxOptions vertxOptions = new VertxOptions();
-            JmxCollectorRegistry jmxCollectorRegistry = null;
-            if (Boolean.parseBoolean(System.getenv(KAFKA_BRIDGE_METRICS_ENABLED))) {
-                LOGGER.info("Metrics enabled and exposed on the /metrics endpoint");
-                // setup Micrometer metrics options
-                vertxOptions.setMetricsOptions(metricsOptions());
-                jmxCollectorRegistry = getJmxCollectorRegistry();
-            }
-            Vertx vertx = Vertx.vertx(vertxOptions);
-            // MeterRegistry default instance is just null if metrics are not enabled in the VertxOptions instance
-            MeterRegistry meterRegistry = BackendRegistries.getDefaultNow();
-            MetricsReporter metricsReporter = new MetricsReporter(jmxCollectorRegistry, meterRegistry);
-
-
             CommandLine commandLine = new DefaultParser().parse(generateOptions(), args);
-
             Map<String, Object> config = ConfigRetriever.getConfig(absoluteFilePath(commandLine.getOptionValue("config-file")));
             BridgeConfig bridgeConfig = BridgeConfig.fromMap(config);
             LOGGER.info("Bridge configuration {}", bridgeConfig);
 
-            deployHttpBridge(vertx, bridgeConfig, metricsReporter).onComplete(done -> {
+            deployHttpBridge(bridgeConfig).onComplete(done -> {
                 if (done.succeeded()) {
                     // register tracing - if set, etc
                     TracingUtil.initialize(bridgeConfig);
@@ -89,6 +75,45 @@ public class Application {
     }
 
     /**
+     * Deploys the HTTP bridge into a new verticle
+     *
+     * @param bridgeConfig          Bridge configuration
+     * @return                      Future for the bridge startup
+     */
+    private static Future<HttpBridge> deployHttpBridge(BridgeConfig bridgeConfig) 
+            throws MalformedObjectNameException, IOException {
+        Promise<HttpBridge> httpPromise = Promise.promise();
+
+        Vertx vertx = createVertxInstance(bridgeConfig);
+        MetricsReporter metricsReporter = getMetricsReporter(bridgeConfig);
+        HttpBridge httpBridge = new HttpBridge(bridgeConfig, metricsReporter);
+        vertx.deployVerticle(httpBridge)
+            .onComplete(done -> {
+                if (done.succeeded()) {
+                    LOGGER.info("HTTP verticle instance deployed [{}]", done.result());
+                    if (metricsReporter != null) {
+                        LOGGER.info("Metrics of type '{}' enabled and exposed on /metrics endpoint", bridgeConfig.getMetrics());
+                    }
+                    httpPromise.complete(httpBridge);
+                } else {
+                    LOGGER.error("Failed to deploy HTTP verticle instance", done.cause());
+                    httpPromise.fail(done.cause());
+                }
+            });
+
+        return httpPromise.future();
+    }
+
+    private static Vertx createVertxInstance(BridgeConfig bridgeConfig) {
+        VertxOptions vertxOptions = new VertxOptions();
+        if (bridgeConfig.getMetrics() != null) {
+            vertxOptions.setMetricsOptions(metricsOptions()); // enable Vertx metrics
+        }
+        Vertx vertx = Vertx.vertx(vertxOptions);
+        return vertx;
+    }
+
+    /**
      * Set up the Vert.x metrics options
      *
      * @return instance of the MicrometerMetricsOptions on Vert.x
@@ -98,60 +123,54 @@ public class Application {
         set.add(MetricsDomain.NAMED_POOLS.name());
         set.add(MetricsDomain.VERTICLES.name());
         return new MicrometerMetricsOptions()
-                .setPrometheusOptions(new VertxPrometheusOptions().setEnabled(true))
-                // define the labels on the HTTP server related metrics
-                .setLabels(EnumSet.of(Label.HTTP_PATH, Label.HTTP_METHOD, Label.HTTP_CODE))
-                // disable metrics about pool and verticles
-                .setDisabledMetricsCategories(set)
-                .setJvmMetricsEnabled(true)
-                .setEnabled(true);
+            .setPrometheusOptions(new VertxPrometheusOptions().setEnabled(true))
+            // define the labels on the HTTP server related metrics
+            .setLabels(EnumSet.of(Label.HTTP_PATH, Label.HTTP_METHOD, Label.HTTP_CODE))
+            // disable metrics about pool and verticles
+            .setDisabledMetricsCategories(set)
+            .setJvmMetricsEnabled(true)
+            .setEnabled(true);
+    }
+
+    private static MetricsReporter getMetricsReporter(BridgeConfig bridgeConfig) 
+            throws MalformedObjectNameException, IOException {
+        if (bridgeConfig.getMetrics() != null) {
+            if (bridgeConfig.getMetrics().equals(MetricsType.JMX_EXPORTER.toString())) {
+                return new MetricsReporter(getJmxCollectorRegistry(bridgeConfig));
+            } else if (bridgeConfig.getMetrics().equals(MetricsType.STRIMZI_REPORTER.toString())) {
+                return new MetricsReporter(new StrimziCollectorRegistry());
+            }
+        }
+        return null;
     }
 
     /**
-     * Deploys the HTTP bridge into a new verticle
-     *
-     * @param vertx                 Vertx instance
-     * @param bridgeConfig          Bridge configuration
-     * @param metricsReporter       MetricsReporter instance for scraping metrics from different registries
-     * @return                      Future for the bridge startup
-     */
-    private static Future<HttpBridge> deployHttpBridge(Vertx vertx, BridgeConfig bridgeConfig, MetricsReporter metricsReporter)  {
-        Promise<HttpBridge> httpPromise = Promise.promise();
-
-        HttpBridge httpBridge = new HttpBridge(bridgeConfig, metricsReporter);
-        vertx.deployVerticle(httpBridge)
-                .onComplete(done -> {
-                    if (done.succeeded()) {
-                        LOGGER.info("HTTP verticle instance deployed [{}]", done.result());
-                        httpPromise.complete(httpBridge);
-                    } else {
-                        LOGGER.error("Failed to deploy HTTP verticle instance", done.cause());
-                        httpPromise.fail(done.cause());
-                    }
-                });
-
-        return httpPromise.future();
-    }
-
-    /**
-     * Return a JmxCollectorRegistry instance with the YAML configuration filters
+     * Return a JmxCollectorRegistry instance with the YAML configuration filters.
+     * This is loaded from a custom config file if present, or from the default configuration file.
      *
      * @return JmxCollectorRegistry instance
      * @throws MalformedObjectNameException
      * @throws IOException
      */
-    private static JmxCollectorRegistry getJmxCollectorRegistry()
-            throws MalformedObjectNameException, IOException {
-        InputStream is = Application.class.getClassLoader().getResourceAsStream("jmx_metrics_config.yaml");
-        if (is == null) {
-            return null;
-        }
-
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8))) {
-            String yaml = reader
+    private static JmxCollectorRegistry getJmxCollectorRegistry(BridgeConfig bridgeConfig) throws MalformedObjectNameException, IOException {
+        if (bridgeConfig.getJmxExporterConfigPath() != null && Files.exists(bridgeConfig.getJmxExporterConfigPath())) {
+            // read custom configuration file
+            LOGGER.info("Loading custom JMX Exporter configuration from {}", bridgeConfig.getJmxExporterConfigPath());
+            String yaml = Files.readString(bridgeConfig.getJmxExporterConfigPath(), StandardCharsets.UTF_8);
+            return new JmxCollectorRegistry(yaml);
+        } else {
+            // fallback to default configuration
+            LOGGER.info("Loading default JMX Exporter configuration");
+            InputStream is = Application.class.getClassLoader().getResourceAsStream("jmx_metrics_config.yaml");
+            if (is == null) {
+                return null;
+            }
+            try (BufferedReader reader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8))) {
+                String yaml = reader
                     .lines()
                     .collect(Collectors.joining("\n"));
-            return new JmxCollectorRegistry(yaml);
+                return new JmxCollectorRegistry(yaml);
+            }
         }
     }
 
@@ -161,7 +180,6 @@ public class Application {
      * @return command line options
      */
     private static Options generateOptions() {
-
         Option configFileOption = Option.builder()
                 .required(true)
                 .hasArg(true)