+++ src/main/java/io/strimzi/kafka/bridge/http/HttpBridge.java
@@ -15,10 +15,10 @@ import io.strimzi.kafka.bridge.BridgeContentType;
 import io.strimzi.kafka.bridge.ConsumerInstanceId;
 import io.strimzi.kafka.bridge.EmbeddedFormat;
 import io.strimzi.kafka.bridge.IllegalEmbeddedFormatException;
-import io.strimzi.kafka.bridge.MetricsReporter;
 import io.strimzi.kafka.bridge.config.BridgeConfig;
 import io.strimzi.kafka.bridge.http.converter.JsonUtils;
 import io.strimzi.kafka.bridge.http.model.HttpBridgeError;
+import io.strimzi.kafka.bridge.metrics.MetricsReporter;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.Promise;
 import io.vertx.core.file.FileSystem;
@@ -140,7 +140,6 @@ public class HttpBridge extends AbstractVerticle {
 
     @Override
     public void start(Promise<Void> startPromise) {
-
         RouterBuilder.create(vertx, "openapi.json")
                 .onComplete(ar -> {
                     if (ar.succeeded()) {
@@ -169,7 +168,9 @@ public class HttpBridge extends AbstractVerticle {
                         routerBuilder.operation(this.OPENAPI.getOperationId().toString()).handler(this.OPENAPI);
                         routerBuilder.operation(this.OPENAPIV2.getOperationId().toString()).handler(this.OPENAPIV2);
                         routerBuilder.operation(this.OPENAPIV3.getOperationId().toString()).handler(this.OPENAPIV3);
-                        routerBuilder.operation(this.METRICS.getOperationId().toString()).handler(this.METRICS);
+                        if (metricsReporter != null) {
+                            routerBuilder.operation(this.METRICS.getOperationId().toString()).handler(this.METRICS);
+                        }
                         routerBuilder.operation(this.INFO.getOperationId().toString()).handler(this.INFO);
                         if (this.bridgeConfig.getHttpConfig().isCorsEnabled()) {
                             routerBuilder.rootHandler(getCorsHandler());
@@ -179,14 +180,14 @@ public class HttpBridge extends AbstractVerticle {
                         }
 
                         this.router = routerBuilder.createRouter();
-
+                        
                         // handling validation errors and not existing endpoints
                         this.router.errorHandler(HttpResponseStatus.BAD_REQUEST.code(), this::errorHandler);
                         this.router.errorHandler(HttpResponseStatus.NOT_FOUND.code(), this::errorHandler);
-
-                        if (this.metricsReporter.getMeterRegistry() != null) {
+        
+                        if (this.metricsReporter != null && this.metricsReporter.getVertxRegistry() != null) {
                             // exclude to report the HTTP server metrics for the /metrics endpoint itself
-                            this.metricsReporter.getMeterRegistry().config().meterFilter(
+                            this.metricsReporter.getVertxRegistry().config().meterFilter(
                                     MeterFilter.deny(meter -> "/metrics".equals(meter.getTag(Label.HTTP_PATH.toString())))
                             );
                         }
@@ -607,6 +608,8 @@ public class HttpBridge extends AbstractVerticle {
             }
         } else if (routingContext.statusCode() == HttpResponseStatus.NOT_FOUND.code()) {
             message = HttpResponseStatus.NOT_FOUND.reasonPhrase();
+        } else if (routingContext.statusCode() == HttpResponseStatus.NOT_IMPLEMENTED.code()) {
+            message = HttpResponseStatus.NOT_IMPLEMENTED.reasonPhrase();
         }
 
         HttpBridgeError error = new HttpBridgeError(routingContext.statusCode(), message);
@@ -823,7 +826,7 @@ public class HttpBridge extends AbstractVerticle {
             openapi(routingContext);
         }
     };
-
+    
     final static HttpOpenApiOperation OPENAPIV2 = new HttpOpenApiOperation(HttpOpenApiOperations.OPENAPIV2) {
 
         @Override