+++ src/test/java/io/strimzi/kafka/bridge/http/base/HttpBridgeITAbstract.java
@@ -5,9 +5,6 @@
 
 package io.strimzi.kafka.bridge.http.base;
 
-import io.micrometer.core.instrument.MeterRegistry;
-import io.strimzi.kafka.bridge.JmxCollectorRegistry;
-import io.strimzi.kafka.bridge.MetricsReporter;
 import io.strimzi.kafka.bridge.clients.BasicKafkaClient;
 import io.strimzi.kafka.bridge.config.BridgeConfig;
 import io.strimzi.kafka.bridge.config.KafkaConfig;
@@ -20,6 +17,9 @@ import io.strimzi.kafka.bridge.http.services.BaseService;
 import io.strimzi.kafka.bridge.http.services.ConsumerService;
 import io.strimzi.kafka.bridge.http.services.ProducerService;
 import io.strimzi.kafka.bridge.http.services.SeekService;
+import io.strimzi.kafka.bridge.metrics.JmxCollectorRegistry;
+import io.strimzi.kafka.bridge.metrics.MetricsReporter;
+import io.strimzi.kafka.bridge.metrics.StrimziCollectorRegistry;
 import io.strimzi.kafka.bridge.utils.Urls;
 import io.strimzi.test.container.StrimziKafkaContainer;
 import io.vertx.core.Vertx;
@@ -97,8 +97,6 @@ public abstract class HttpBridgeITAbstract {
     protected static AdminClientFacade adminClientFacade;
     protected static HttpBridge httpBridge;
     protected static BridgeConfig bridgeConfig;
-
-    protected static MeterRegistry meterRegistry = null;
     protected static JmxCollectorRegistry jmxCollectorRegistry = null;
 
     protected BaseService baseService() {
@@ -128,7 +126,8 @@ public abstract class HttpBridgeITAbstract {
 
         if ("FALSE".equals(BRIDGE_EXTERNAL_ENV)) {
             bridgeConfig = BridgeConfig.fromMap(config);
-            httpBridge = new HttpBridge(bridgeConfig, new MetricsReporter(jmxCollectorRegistry, meterRegistry));
+            
+            httpBridge = new HttpBridge(bridgeConfig, new MetricsReporter(new StrimziCollectorRegistry()));
 
             LOGGER.info("Deploying in-memory bridge");
             vertx.deployVerticle(httpBridge).onComplete(context.succeeding(id -> context.completeNow()));